BIB-444: Consolidated sip2 fields on sip2user entity
