+++ src/Controller/Admin/SchoolCrudController.php
@@ -37,7 +37,6 @@ class SchoolCrudController extends AbstractCrudController
         return [
             IdField::new('id')->hideOnForm()->hideOnIndex(),
             TextField::new('name'),
-            TextField::new('sip2InstitutionId'),
             AssociationField::new('boxConfigurations'),
         ];
     }
