+++ src/DataFixtures/AppFixtures.php
@@ -79,8 +79,6 @@ class AppFixtures extends Fixture
         foreach (self::$schoolNames as $schoolName) {
             $school = new School();
             $school->setName($schoolName);
-            $libraryNumber = $faker->numberBetween(70000, 90000) * 10;
-            $school->setSip2InstitutionId('DK-'.$libraryNumber);
 
             $schools[] = $school;
             $manager->persist($school);
@@ -90,6 +88,8 @@ class AppFixtures extends Fixture
             $sip2user = new Sip2User();
             $sip2user->setUsername($faker->userName);
             $sip2user->setPassword($faker->password);
+            $libraryNumber = $faker->numberBetween(70000, 90000) * 10;
+            $sip2user->setSip2InstitutionId('DK-'.$libraryNumber);
 
             $sip2users[] = $sip2user;
             $manager->persist($sip2user);
